+++ src/learning/nash_q_learning.cpp
@@ -3,19 +3,19 @@
 namespace perimeter
 {
 
-NashQLearning::NashQLearning(int id, int numAgents, double gamma,
+NashQLearning::NashQLearning(int id, int numAgents, float gamma,
                              const JointActionSpace& jointActionSpace, core::AgentType agentType)
     : id_{id}
     , agentType_{agentType}
     , numAgents_{numAgents}
     , gamma_{gamma}
-    , randomPolicy_{std::vector<double>(static_cast<int>(environment::Action::NUM_ACTIONS),
-                                        1 / static_cast<double>(environment::Action::NUM_ACTIONS))}
+    , randomPolicy_{std::vector<float>(static_cast<int>(environment::Action::NUM_ACTIONS),
+                                        1 / static_cast<float>(environment::Action::NUM_ACTIONS))}
     , policy_{randomPolicy_}
     , jointActionSpace_{jointActionSpace}
 {}
 
-const std::function<double(JointState, JointAction)> NashQLearning::getRewardFunction()
+const std::function<float(JointState, JointAction)> NashQLearning::getRewardFunction()
 {
   return [this](const JointState& state, const JointAction& action) { return Q(state, action); };
 }
@@ -34,14 +34,14 @@ void NashQLearning::updateJointQTable(const JointState& prevAgentStates,
   const JointState& s = prevAgentStates;
   const JointState& s_prime = currAgentStates;
   const JointAction& a = prevJointAction;
-  double reward = stepRewards.at(id_);
+  float reward = stepRewards.at(id_);
 
-  double utility{0.0};
+  float utility{0.0};
   for (const JointAction& a_prime : jointActionSpace_.getJointActionSpace()) {
     utility += Q(s_prime, a_prime) * jointActionProbability(jointPolicy, a_prime);
   }
 
-  double alpha = 1 / std::sqrt(static_cast<double>(N(s, a)));
+  float alpha = 1 / std::sqrt(static_cast<float>(N(s, a)));
   Q_s_a_[s][a] = Q(s, a) + alpha * (reward + gamma_ * utility - Q(s, a));
 }
 
@@ -52,7 +52,7 @@ const environment::Action NashQLearning::sampleEpsGreedyPolicy(std::mt19937& rg,
     N_s_[state] = 1;
   }
 
-  double eps = 1 / N_s_[state];
+  float eps = 1 / N_s_[state];
   std::bernoulli_distribution dist(eps);
   if (dist(rg)) {
     return randomPolicy_.sampleAction(rg);
@@ -73,7 +73,7 @@ void NashQLearning::updateN(const JointState& state, const JointAction& jointAct
   N_s_[state] += 1;
 }
 
-double NashQLearning::Q(const JointState& state, const JointAction& jointAction)
+float NashQLearning::Q(const JointState& state, const JointAction& jointAction)
 {
   if (!Q_s_a_.contains(state) || !Q_s_a_[state].contains(jointAction)) {
     Q_s_a_[state][jointAction] = computeInitialQTableValue(state, jointAction);
@@ -89,20 +89,20 @@ int NashQLearning::N(const JointState& state, const JointAction& jointAction)
   return N_s_a_[state][jointAction];
 }
 
-double NashQLearning::computeInitialQTableValue(const JointState& state,
+float NashQLearning::computeInitialQTableValue(const JointState& state,
                                                 const JointAction& jointAction) const
 {
   // Bootstrap your Q table here
   return 0.0;
 }
 
-double jointActionProbability(const JointPolicy& jointPolicy, const JointAction& jointAction)
+float jointActionProbability(const JointPolicy& jointPolicy, const JointAction& jointAction)
 {
   if (jointPolicy.size() != jointAction.size()) {
     throw std::invalid_argument("Joint policy must be the same size as the joint action");
   }
 
-  double product{1.0};
+  float product{1.0};
   for (int i{0}; i < jointPolicy.size(); ++i) {
     const SingleAgentSimpleGamePolicy& policyI = jointPolicy[i];
     product *= policyI.getProbability(jointAction[i]);
